Pull checkBinary up to top level
