+++ ctags/json.go
@@ -15,12 +15,9 @@
 package ctags
 
 import (
-	"bytes"
 	"fmt"
 	"log"
 	"os"
-	"os/exec"
-	"strings"
 	"sync"
 	"time"
 
@@ -121,11 +118,7 @@ func (lp *lockedParser) close() {
 
 // NewParser creates a parser that is implemented by the given
 // universal-ctags binary. The parser is safe for concurrent use.
-func NewParser(parserType CTagsParserType, bin string) (Parser, error) {
-	if err := checkBinary(parserType, bin); err != nil {
-		return nil, err
-	}
-
+func NewParser(bin string) Parser {
 	opts := goctags.Options{
 		Bin: bin,
 	}
@@ -133,29 +126,5 @@ func NewParser(parserType CTagsParserType, bin string) (Parser, error) {
 		opts.Info = log.New(os.Stderr, "CTAGS INF: ", log.LstdFlags)
 		opts.Debug = log.New(os.Stderr, "CTAGS DBG: ", log.LstdFlags)
 	}
-	return &lockedParser{
-		opts: opts,
-	}, nil
-}
-
-// checkBinary does checks on bin to ensure we can correctly use the binary
-// for symbols. It is more user friendly to fail early in this case.
-func checkBinary(typ CTagsParserType, bin string) error {
-	switch typ {
-	case UniversalCTags:
-		helpOutput, err := exec.Command(bin, "--help").CombinedOutput()
-		if err != nil {
-			return fmt.Errorf("failed to check if %s is universal-ctags: %w\n--help output:\n%s", bin, err, string(helpOutput))
-		}
-		if !bytes.Contains(helpOutput, []byte("+interactive")) {
-			return fmt.Errorf("ctags binary is not universal-ctags or is not compiled with +interactive feature: bin=%s", bin)
-		}
-
-	case ScipCTags:
-		if !strings.Contains(bin, "scip-ctags") {
-			return fmt.Errorf("only supports scip-ctags, not %s", bin)
-		}
-	}
-
-	return nil
+	return &lockedParser{opts: opts,}
 }
