+++ ctags/json_test.go
@@ -27,11 +27,7 @@ func TestJSON(t *testing.T) {
 		t.Skip(err)
 	}
 
-	p, err := NewParser(UniversalCTags, "universal-ctags")
-	if err != nil {
-		t.Fatal("newProcess", err)
-	}
-
+	p := NewParser("universal-ctags")
 	defer p.Close()
 
 	java := `
